--- DSharpPlus.Commands/Attributes/FromCode/CodeType.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace DSharpPlus.Commands.Attributes;
-
-[Flags]
-public enum CodeType
-{
-    Inline = 1 << 0,
-    Codeblock = 1 << 1,
-    All = Inline | Codeblock
-}
